requirements.txt
hydra-core
loguru
omegaconf
pytorch-lightning
pytorch-lightning-bolts
scikit_learn
torch
torchvision
